requirements.txt
discord
PyYAML
requests
pillow
pytz
apscheduler
